+++ tests/test_countdown.py
@@ -14,6 +14,10 @@ class CountdownTest(TestCase):
         self.assertTrue(test_cd.timed_out())
         self.assertTrue(test_cd.rem_time() == timedelta())
         test_cd.timeout = timedelta(seconds=0.1)
+        self.assertEqual(
+            test_cd.timeout.total_seconds(), timedelta(seconds=0.1).total_seconds()
+        )
+        self.assertEqual(test_cd.timeout_ms, 100)
         self.assertTrue(test_cd.busy())
         self.assertFalse(test_cd.timed_out())
         time.sleep(0.1)