+++ tmtccmd/cfdp/handler/source.py
@@ -312,7 +312,19 @@ class SourceHandler:
 
     def cancel_request(self, transaction_id: TransactionId) -> bool:
         """This function models the Cancel.request CFDP primtive and is the recommended way
-        to cancel a transaction. It will cause a Notice Of Cancellation at this entity."""
+        to cancel a transaction. It will cause a Notice Of Cancellation at this entity.
+        Please note that the state machine might still be active because a canceled transfer
+        might still require some packets to be sent to the remote receiver entity.
+
+        Returns
+        --------
+        True
+            Current transfer was cancelled
+        False
+            The state machine is in the IDLE state or there is a transaction ID missmatch.
+        """
+        if self.states.step == CfdpState.IDLE:
+            return False
         if self.states.packets_ready:
             raise UnretrievedPdusToBeSent()
         if (